pyproject.toml
[project]
name = "mcp-server-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-api-python-client>=2.188.0",
    "google-auth>=2.48.0",
    "google-auth-oauthlib>=1.2.4",
    "mcp[cli]>=1.25.0",
    "python-dateutil>=2.9.0.post0",
]
